Refactor a bunch
