+++ src/controllers/api/raceController.js
@@ -2,8 +2,7 @@ const Race = require('../../models/race');
 const Subrace = require('../../models/subrace');
 const Trait = require('../../models/trait');
 const Proficiency = require('../../models/proficiency');
-
-const utility = require('./utility');
+const { ResourceList } = require('../../util');
 const SimpleController = require('../simpleController');
 
 const simpleController = new SimpleController(Race);
@@ -16,7 +15,7 @@ exports.showSubracesForRace = (req, res, next) => {
   return Subrace.find({ 'race.url': urlString })
     .select({ index: 1, name: 1, url: 1, _id: 0 })
     .then(data => {
-      res.status(200).json(utility.ResourceList(data));
+      res.status(200).json(ResourceList(data));
     })
     .catch(err => {
       next(err);
@@ -28,7 +27,7 @@ exports.showTraitsForRace = (req, res, next) => {
   return Trait.find({ 'races.url': urlString })
     .select({ index: 1, name: 1, url: 1, _id: 0 })
     .then(data => {
-      res.status(200).json(utility.ResourceList(data));
+      res.status(200).json(ResourceList(data));
     })
     .catch(err => {
       next(err);
@@ -42,7 +41,7 @@ exports.showProficienciesForRace = (req, res, next) => {
     .select({ index: 1, name: 1, url: 1, _id: 0 })
     .sort({ index: 'asc' })
     .then(data => {
-      res.status(200).json(utility.ResourceList(data));
+      res.status(200).json(ResourceList(data));
     })
     .catch(err => {
       next(err);
