+++ src/controllers/api/spellController.js
@@ -1,13 +1,12 @@
 const { promisify } = require('util');
 const Spell = require('../../models/spell');
-const utility = require('./utility');
-const { redisClient } = require('../../util');
+const { redisClient, escapeRegExp, ResourceList } = require('../../util');
 const getAsync = promisify(redisClient.get).bind(redisClient);
 
 exports.index = async (req, res, next) => {
   const searchQueries = {};
   if (req.query.name !== undefined) {
-    searchQueries.name = { $regex: new RegExp(utility.escapeRegExp(req.query.name), 'i') };
+    searchQueries.name = { $regex: new RegExp(escapeRegExp(req.query.name), 'i') };
   }
 
   if (req.query.level !== undefined) {
@@ -31,7 +30,7 @@ exports.index = async (req, res, next) => {
       .select({ index: 1, name: 1, url: 1, _id: 0 })
       .sort({ index: 'asc' })
       .then(data => {
-        const jsonData = utility.ResourceList(data);
+        const jsonData = ResourceList(data);
         redisClient.set(redisKey, JSON.stringify(jsonData));
         res.status(200).json(jsonData);
       })
