+++ src/util/regex.js
@@ -1,0 +1,7 @@
+const escapeRegExp = string => {
+  return string.toString().replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& means the whole matched string
+};
+
+module.exports = {
+  escapeRegExp
+};
